+++ ejercicios/project/entropy.cpp
@@ -7,7 +7,7 @@ double entropy(std::vector<double> P)
     
     /*es halla la entropía sumando los elementos P*ln(P). se descartan los
       P=0*/
-    for(int i=0;i<P.size();i++){
+    for(long unsigned int i=0;i<P.size();i++){
         if(P[i]!=0){
             E+=P[i]*std::log(P[i]);
         }
