+++ ejercicios/project/n_state.cpp
@@ -9,7 +9,7 @@ void n_state(std::vector<double> &r,double l,std::mt19937 &gen)
     /*a cada elemento del vector, se suma una cantidad aleatoria pequeña
       y se verifica que el nuevo elemento no se salga de las paredes, en
       tal caso, se le devuelve a la pared*/
-    for(int i=0;i<r.size();i++){
+    for(long unsigned int i=0;i<r.size();i++){
         r[i]+=dis(gen);
         if(r[i]>l){
             r[i]=l;
