+++ ejercicios/project/print.cpp
@@ -3,7 +3,7 @@
 
 void print(std::vector<double> r)
 {
-    for(int i=0;i<r.size();i+=2){
+    for(long unsigned int i=0;i<r.size();i+=2){
         std::cout<<r[i]<<"\t"<<r[i+1]<<"\n";
     }
 }
